requirements.txt
# Streaming
kafka-python==2.0.2
confluent-kafka==2.3.0

# Feature Store
redis==5.0.1
hiredis==2.2.3

# ML
numpy==1.26.0
scikit-learn==1.3.2
xgboost==2.0.3
onnxruntime==1.16.3

# Config & Utils
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
